Fixing some errors

+++ app/Observers/ProductObserver.php
@@ -20,6 +20,14 @@ class ProductObserver
      */
     public function saved(Product $product): void
     {
+        //
+    }
+
+    /**
+     * Handle the Product "updated" event.
+     */
+    public function updated(Product $product): void
+    {
         if ($product->arrayAttributeIsDirty($product->image_urls)) {
             foreach ($product->dirtyArrayElements($product->image_urls) as $image_url) {
                 Storage::disk('public')->delete($image_url);
@@ -28,19 +36,11 @@ class ProductObserver
     }
 
     /**
-     * Handle the Product "updated" event.
-     */
-    public function updated(Product $product): void
-    {
-        //
-    }
-
-    /**
      * Handle the Product "deleted" event.
      */
     public function deleted(Product $product): void
     {
-        if (!is_null($product->image_urls) || $product->image_urls === []) {
+        if (!is_null($product->image_urls) && count($product->image_urls) > 0) {
             foreach ($product->dirtyArrayElements($product->image_urls) as $image_url) {
                 Storage::disk('public')->delete($image_url);
             }
